allow auth for talk client
extract common client code to module and allow auth server to be configurable for talk API
